backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.0",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "firecrawl-py>=4.8.0",
    "geopy>=2.4.1",
    "google-generativeai>=0.8.3",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.8",
    "langchain==0.3.27",
    "langgraph>=0.2.59",
    "langchain-community>=0.3.14",
    "langchain-qdrant>=0.2.0",
    "qdrant-client>=1.12.1",
    "opencage>=3.2.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
    "websockets>=13.1",
    "uvicorn>=0.38.0",
]
